Created Admin page for leaves
